Consolidate artifact iframe HTML across r/, reports/ and MCP
The standard report viewer (pages/r/[id]/index.vue) and the in-app
editor (components/dashboard/ArtifactFrame.vue) each built a near-
identical srcdoc — same React/Babel/ECharts/Tailwind scripts, same
artifact-globals.js inclusion — by copy-paste. Meanwhile the MCP
iframe at frontend/public/mcp-artifact-app.html (loaded by Claude
Desktop / Cursor over the MCP Apps postMessage transport) had its
own runtime that exposed React but not the bare-name hooks
(useMemo, useState, …) that the LLM prompt at
backend/app/ai/tools/implementations/_sandbox_context.py instructs
the model to use, so MCP-rendered artifacts crashed with
"useMemo is not defined" while the same code rendered correctly in
the web app.

Extract a single buildArtifactIframeHtml() helper used by both Vue
pages, and add artifact-globals.js to the MCP HTML so its runtime
surface matches. The MCP runtime block still overrides
useArtifactData with an event-driven variant because MCP delivers
ARTIFACT_DATA over postMessage rather than inlined srcdoc.
